Less modules, less functions, m1download_utils.py can be removed
